requirements.txt
fastapi
uvicorn
pytest
python-dotenv==1.0.0
pydantic[email]
httpx